Refactor:added size prop to Button and fixd imports in some routes
