+++ app/routes/equipo/HeroEquipo.jsx
@@ -1,4 +1,4 @@
-import ButtonLink from "../../components/Button/ButtonLink";
+import ButtonLink from "../../components/Buttons/ButtonLink";
 
 export default function HeroEquipo() {
   return (
